change length to return int

+++ vector.h
@@ -17,7 +17,7 @@ void vector_delete(Vector*);
 
 void* vector_get(Vector*, size_t);
 int vector_in(Vector*, void*);
-size_t vector_length(Vector*);
+int vector_length(Vector*);
 void* vector_pop(Vector*);
 void vector_push(Vector*, void*);
 void vector_reverse(Vector*);
